api/pyproject.toml
[project]
name = "api-python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0",
  "bcrypt>=5.0.0",
  "cachetools>=5.3.0",
  "fastapi>=0.124.2",
  "firebase-admin>=6.0.0",
  "httpx>=0.28.1",
  "langchain-openai>=0.3.0",
  "langchain-community>=0.3.0",
  "langgraph>=0.2.0",
  "psycopg2-binary>=2.9.11",
  "pydantic-settings>=2.12.0",
  "pydantic[email]>=2.12.5",
  "python-dotenv>=1.2.1",
  "python-jose[cryptography]>=3.5.0",
  "sqlalchemy>=2.0.45",
  "stripe>=14.1.0",
  "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = ["ruff>=0.14.8", "pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.28.1", "passlib[bcrypt]>=1.7.4"]
